Monkey patch FFI::MemoryPointer to add from_data.
There are many places in the code base where we create an
FFI::MemoryPointer buffer and then call write_bytes immediately,
before passing it on to an FFI call.
It's easy to forget the second step, and this can have serious
consequences (bug fix incoming).

FFI::MemoryPointer has from_string but this creates a buffer
of size (len+1) for the null terminator, and we tend to rely on
FFI::MemoryPointer#size for this information.